Add TryFromFeltLibfunc.
